refactor(decomp): move consolidated_reader to proximadb-raptor-engine crate (first raptor leaf)
Move `src/storage/engines/raptor/consolidated_reader.rs` (+ its sibling
`consolidated_reader/tests.rs`) into a new `proximadb-raptor-engine` crate — the
first raptor engine leaf extracted out of the root crate (the direct OOM-payoff
step all the enabler work built toward). Root `consolidated_reader.rs` becomes a
one-line re-export shim, so every `crate::storage::engines::raptor::consolidated_reader::*`
path (RaptorReader, ScanStrategy, IntraRowgroupMatrix — the 3 consumers: engine.rs,
consolidated_compactor.rs, extraction.rs, mod.rs) resolves unchanged.

This is the consolidated_reader extraction (commit 2/2, stacked on the #958
config/constants hoist + the TC-removal commit 1/2). Its deps are all crate-accessible
after the prior enablers:
  - common/config/constants → proximadb-raptor-common (#955, #958)
  - cache → Arc<dyn CacheAccessPatternPort> (#948)
  - FileSystem trait → proximadb-storage-filesystem-types
  - proto/search-types/codec → proximadb-proto/search-types/codec (free-swaps)
  - TransactionCoordinator → dead param removed (commit 1/2)

Mechanical: history-preserving `git mv` + 7 import free-swaps + the sibling tests.rs
moved into the crate (its `super::super::common`→`proximadb_raptor_common`).
`experimental-engines` feature declared in the crate + propagated from root (the
tests mod was already experimental-gated; behavior preserved).

~4,176 LOC out of the root single-compile unit. Storage up-edge ratchet 181 (held —
consolidated_reader's proto/core refs weren't up-edge targets).

Verified: root --lib + raptor-engine crate + clippy --lib --bins + crate --tests
all clean; fmt clean; layering + workspace-boundaries pass; work-commit-check green.
