+++ crates/storage/proximadb-raptor-engine/Cargo.toml
@@ -1,0 +1,43 @@
+[package]
+name = "proximadb-raptor-engine"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "RAPTOR engine — consolidated reader (the first raptor leaf moved out of the root crate); depends on proximadb-raptor-common downward. Root-crate decomposition (engines extraction)."
+
+[lib]
+name = "proximadb_raptor_engine"
+path = "src/lib.rs"
+
+[dependencies]
+anyhow.workspace = true
+chrono.workspace = true
+serde.workspace = true
+async-trait.workspace = true
+tokio.workspace = true
+dashmap.workspace = true
+bincode.workspace = true
+serde_json.workspace = true
+tracing.workspace = true
+arrow-array.workspace = true
+arrow-ipc = { version = "58", default-features = false, features = ["lz4", "zstd"] }
+arrow-schema.workspace = true
+proximadb-raptor-common = { path = "../proximadb-raptor-common" }
+proximadb-storage-ports.workspace = true
+proximadb-storage-filesystem-types.workspace = true
+proximadb-proto.workspace = true
+proximadb-search-types.workspace = true
+proximadb-codec.workspace = true
+proximadb-compression.workspace = true
+proximadb-distance-kernel.workspace = true
+proximadb-kernel.workspace = true
+
+[features]
+experimental-engines = []
+
+[lints]
+workspace = true
